+++ packages/isaacscript-common/src/classes/features/other/ModdedElementDetection.ts
@@ -7,9 +7,11 @@ import type {
 import { ModCallback } from "isaac-typescript-definitions";
 import { itemConfig } from "../../../core/cachedClasses";
 import {
-  LAST_VANILLA_CARD_TYPE,
+  FIRST_MODDED_CARD_TYPE,
+  FIRST_MODDED_COLLECTIBLE_TYPE,
+  FIRST_MODDED_PILL_EFFECT,
+  FIRST_MODDED_TRINKET_TYPE,
   LAST_VANILLA_COLLECTIBLE_TYPE,
-  LAST_VANILLA_PILL_EFFECT,
   LAST_VANILLA_TRINKET_TYPE,
   NUM_VANILLA_CARD_TYPES,
   NUM_VANILLA_COLLECTIBLE_TYPES,
@@ -17,13 +19,6 @@ import {
   NUM_VANILLA_TRINKET_TYPES,
 } from "../../../core/constantsFirstLast";
 import { Exported } from "../../../decorators";
-import {
-  asCardType,
-  asCollectibleType,
-  asNumber,
-  asPillEffect,
-  asTrinketType,
-} from "../../../functions/types";
 import { Feature } from "../../private/Feature";
 
 /**
@@ -82,16 +77,13 @@ export class ModdedElementDetection extends Feature {
   public getFirstModdedCollectibleType(): CollectibleType | undefined {
     this.errorIfNoCallbacksFired("collectible");
 
-    const firstModdedCollectibleType = asCollectibleType(
-      asNumber(LAST_VANILLA_COLLECTIBLE_TYPE) + 1,
-    );
     const itemConfigItem = itemConfig.GetCollectible(
-      firstModdedCollectibleType,
+      FIRST_MODDED_COLLECTIBLE_TYPE,
     );
 
     return itemConfigItem === undefined
       ? undefined
-      : firstModdedCollectibleType;
+      : FIRST_MODDED_COLLECTIBLE_TYPE;
   }
 
   /**
@@ -100,7 +92,7 @@ export class ModdedElementDetection extends Feature {
    * Equal to `itemConfig.GetCollectibles().Size - 1`. (`Size` includes invalid collectibles, like
    * 666. We subtract one to account for `CollectibleType.NULL`.)
    *
-   * If there are no mods present that add any custom items, this function will return
+   * If there are no mods present that add any custom collectibles, this function will return
    * `CollectibleType.MOMS_RING` (732).
    *
    * This function can only be called if at least one callback has been executed. This is because
@@ -120,7 +112,7 @@ export class ModdedElementDetection extends Feature {
 
   /**
    * Returns the total number of collectibles in the item config, including both vanilla and modded
-   * items. If you just need the number of vanilla collectible types, use the
+   * collectibles. If you just need the number of vanilla collectible types, use the
    * `NUM_VANILLA_COLLECTIBLE_TYPES` constant.
    *
    * This function can only be called if at least one callback has been executed. This is because
@@ -135,9 +127,7 @@ export class ModdedElementDetection extends Feature {
   @Exported
   public getNumCollectibleTypes(): int {
     this.errorIfNoCallbacksFired("collectible");
-
-    const numModdedCollectibleTypes = this.getNumModdedCollectibleTypes();
-    return NUM_VANILLA_COLLECTIBLE_TYPES + numModdedCollectibleTypes;
+    return NUM_VANILLA_COLLECTIBLE_TYPES + this.getNumModdedCollectibleTypes();
   }
 
   /**
@@ -155,9 +145,7 @@ export class ModdedElementDetection extends Feature {
   @Exported
   public getNumModdedCollectibleTypes(): int {
     this.errorIfNoCallbacksFired("collectible");
-
-    const lastCollectibleType = this.getLastCollectibleType();
-    return lastCollectibleType - LAST_VANILLA_COLLECTIBLE_TYPE;
+    return this.getLastCollectibleType() - LAST_VANILLA_COLLECTIBLE_TYPE;
   }
 
   // --------
@@ -179,11 +167,8 @@ export class ModdedElementDetection extends Feature {
   public getFirstModdedTrinketType(): TrinketType | undefined {
     this.errorIfNoCallbacksFired("trinket");
 
-    const firstModdedTrinketType = asTrinketType(
-      asNumber(LAST_VANILLA_TRINKET_TYPE) + 1,
-    );
-    const itemConfigItem = itemConfig.GetTrinket(firstModdedTrinketType);
-    return itemConfigItem === undefined ? undefined : firstModdedTrinketType;
+    const itemConfigItem = itemConfig.GetTrinket(FIRST_MODDED_TRINKET_TYPE);
+    return itemConfigItem === undefined ? undefined : FIRST_MODDED_TRINKET_TYPE;
   }
 
   /**
@@ -211,7 +196,7 @@ export class ModdedElementDetection extends Feature {
 
   /**
    * Returns the total number of trinkets in the item config, including both vanilla and modded
-   * items. If you just need the number of vanilla trinket types, use the
+   * trinkets. If you just need the number of vanilla trinket types, use the
    * `NUM_VANILLA_TRINKET_TYPES` constant.
    *
    * This function can only be called if at least one callback has been executed. This is because
@@ -225,9 +210,7 @@ export class ModdedElementDetection extends Feature {
   @Exported
   public getNumTrinketTypes(): int {
     this.errorIfNoCallbacksFired("trinket");
-
-    const numModdedTrinketTypes = this.getNumModdedTrinketTypes();
-    return NUM_VANILLA_TRINKET_TYPES + numModdedTrinketTypes;
+    return NUM_VANILLA_TRINKET_TYPES + this.getNumModdedTrinketTypes();
   }
 
   /**
@@ -244,9 +227,7 @@ export class ModdedElementDetection extends Feature {
   @Exported
   public getNumModdedTrinketTypes(): int {
     this.errorIfNoCallbacksFired("trinket");
-
-    const lastTrinketType = this.getLastTrinketType();
-    return lastTrinketType - LAST_VANILLA_TRINKET_TYPE;
+    return this.getLastTrinketType() - LAST_VANILLA_TRINKET_TYPE;
   }
 
   // -----
@@ -268,22 +249,21 @@ export class ModdedElementDetection extends Feature {
   public getFirstModdedCardType(): CardType | undefined {
     this.errorIfNoCallbacksFired("card");
 
-    const firstModdedCardType = asCardType(
-      asNumber(LAST_VANILLA_CARD_TYPE) + 1,
-    );
-
-    const itemConfigCard = itemConfig.GetCard(firstModdedCardType);
-    return itemConfigCard === undefined ? undefined : firstModdedCardType;
+    const itemConfigCard = itemConfig.GetCard(FIRST_MODDED_CARD_TYPE);
+    return itemConfigCard === undefined ? undefined : FIRST_MODDED_CARD_TYPE;
   }
 
   /**
    * Will change depending on how many modded cards there are.
    *
-   * This is equal to the number of cards, since all card sub-types are contiguous (unlike
-   * collectibles).
+   * Equal to `itemConfig.GetCards().Size - 1`. (`Size` includes invalid cards, but since cards are
+   * contiguous, there are no invalid cards. We subtract one to account for `CardType.NULL`.)
+   *
+   * If there are no mods present that add any custom cards, this function will return
+   * `CardType.SOUL_OF_JACOB_AND_ESAU` (97).
    *
    * This function can only be called if at least one callback has been executed. This is because
-   * not all cards will necessarily be present when a mod first loads (due to mod load order).
+   * not all trinkets will necessarily be present when a mod first loads (due to mod load order).
    *
    * In order to use this function, you must upgrade your mod with
    * `ISCFeature.MODDED_ELEMENT_DETECTION`.
@@ -293,15 +273,12 @@ export class ModdedElementDetection extends Feature {
   @Exported
   public getLastCardType(): CardType {
     this.errorIfNoCallbacksFired("card");
-
-    const numCards = this.getNumCardTypes();
-    return asCardType(numCards);
+    return itemConfig.GetCards().Size - 1;
   }
 
   /**
-   * Will change depending on how many modded cards there are.
-   *
-   * Equal to `itemConfig.GetCards().Size - 1`. (We subtract one to account for `Card.NULL`.)
+   * Returns the total number of cards in the item config, including both vanilla and modded cards.
+   * If you just need the number of vanilla card types, use the `NUM_VANILLA_CARD_TYPES` constant.
    *
    * This function can only be called if at least one callback has been executed. This is because
    * not all cards will necessarily be present when a mod first loads (due to mod load order).
@@ -314,12 +291,14 @@ export class ModdedElementDetection extends Feature {
   @Exported
   public getNumCardTypes(): int {
     this.errorIfNoCallbacksFired("card");
-    return itemConfig.GetCards().Size - 1;
+    return NUM_VANILLA_CARD_TYPES + this.getNumModdedCardTypes();
   }
 
   /**
+   * Like vanilla card types, modded card types are always contiguous.
+   *
    * This function can only be called if at least one callback has been executed. This is because
-   * not all trinkets will necessarily be present when a mod first loads (due to mod load order).
+   * not all cards will necessarily be present when a mod first loads (due to mod load order).
    *
    * In order to use this function, you must upgrade your mod with
    * `ISCFeature.MODDED_ELEMENT_DETECTION`.
@@ -329,9 +308,7 @@ export class ModdedElementDetection extends Feature {
   @Exported
   public getNumModdedCardTypes(): int {
     this.errorIfNoCallbacksFired("card");
-
-    const numCardTypes = this.getNumCardTypes();
-    return numCardTypes - NUM_VANILLA_CARD_TYPES;
+    return this.getLastCardType() - LAST_VANILLA_COLLECTIBLE_TYPE;
   }
 
   // ------------
@@ -354,23 +331,24 @@ export class ModdedElementDetection extends Feature {
   public getFirstModdedPillEffect(): PillEffect | undefined {
     this.errorIfNoCallbacksFired("pill");
 
-    const firstModdedPillEffect = asPillEffect(
-      asNumber(LAST_VANILLA_PILL_EFFECT) + 1,
-    );
     const itemConfigPillEffect = itemConfig.GetPillEffect(
-      firstModdedPillEffect,
+      FIRST_MODDED_PILL_EFFECT,
     );
 
     return itemConfigPillEffect === undefined
       ? undefined
-      : firstModdedPillEffect;
+      : FIRST_MODDED_PILL_EFFECT;
   }
 
   /**
    * Will change depending on how many modded pill effects there are.
    *
-   * This is equal to the number of pill effects, since all pill effects are contiguous (unlike
-   * collectibles).
+   * Equal to `itemConfig.GetPillEffects().Size - 1`. (`Size` includes invalid pill effects, but
+   * since pill effects are contiguous, there are no invalid pill effects. We subtract one to
+   * account for the enum starting at 0 instead of 1.)
+   *
+   * If there are no mods present that add any custom pill effects, this function will return
+   * `PillEffect.EXPERIMENTAL` (49).
    *
    * This function can only be called if at least one callback has been executed. This is because
    * not all pill effects will necessarily be present when a mod first loads (due to mod load
@@ -384,22 +362,16 @@ export class ModdedElementDetection extends Feature {
   @Exported
   public getLastPillEffect(): PillEffect {
     this.errorIfNoCallbacksFired("pill");
-
-    // We need to subtract one because Bad Gas is ID 0. In other words, in vanilla Repentance, there
-    // are 50 pill effects that have IDs from 0 to 49, so the highest pill effect is 49.
-    const numPillEffects = this.getNumPillEffects() - 1;
-    return asPillEffect(numPillEffects);
+    return itemConfig.GetPillEffects().Size - 1;
   }
 
   /**
-   * Will change depending on how many modded pill effects there are.
-   *
-   * Equal to `itemConfig.GetPillEffects().Size`. (We do not have to subtract one, because the first
-   * pill effect has an ID of 0 and there is no `PillEffect.NULL`.)
+   * Returns the total number of pill effects in the item config, including both vanilla and modded
+   * pill effects. If you just need the number of vanilla pill effects, use the
+   * `NUM_VANILLA_PILL_EFFECTS` constant.
    *
    * This function can only be called if at least one callback has been executed. This is because
-   * not all pill effects will necessarily be present when a mod first loads (due to mod load
-   * order).
+   * not all cards will necessarily be present when a mod first loads (due to mod load order).
    *
    * In order to use this function, you must upgrade your mod with
    * `ISCFeature.MODDED_ELEMENT_DETECTION`.
@@ -409,13 +381,14 @@ export class ModdedElementDetection extends Feature {
   @Exported
   public getNumPillEffects(): int {
     this.errorIfNoCallbacksFired("pill");
-    return itemConfig.GetPillEffects().Size;
+    return NUM_VANILLA_PILL_EFFECTS + this.getNumModdedPillEffects();
   }
 
   /**
+   * Like vanilla pill effects, modded pill effects are always contiguous.
+   *
    * This function can only be called if at least one callback has been executed. This is because
-   * not all pill effects will necessarily be present when a mod first loads (due to mod load
-   * order).
+   * not all cards will necessarily be present when a mod first loads (due to mod load order).
    *
    * In order to use this function, you must upgrade your mod with
    * `ISCFeature.MODDED_ELEMENT_DETECTION`.
@@ -424,9 +397,7 @@ export class ModdedElementDetection extends Feature {
    */
   @Exported
   public getNumModdedPillEffects(): int {
-    this.errorIfNoCallbacksFired("pill");
-
-    const numPillEffects = this.getNumPillEffects();
-    return numPillEffects - NUM_VANILLA_PILL_EFFECTS;
+    this.errorIfNoCallbacksFired("card");
+    return this.getLastPillEffect() - LAST_VANILLA_COLLECTIBLE_TYPE;
   }
 }