+++ packages/isaacscript-common/src/classes/features/other/ModdedElementSets.ts
@@ -40,7 +40,6 @@ import { getFlagName } from "../../../functions/flag";
 import { getRandomSetElement } from "../../../functions/set";
 import { trinketHasCacheFlag } from "../../../functions/trinkets";
 import {
-  asCardType,
   asCollectibleType,
   asPillEffect,
   asTrinketType,
@@ -270,7 +269,8 @@ export class ModdedElementSets extends Feature {
 
     for (const cardTypeInt of moddedCardTypes) {
       // Modded card types are contiguous, but we check every value just in case.
-      const cardType = asCardType(cardTypeInt);
+      // eslint-disable-next-line isaacscript/strict-enums
+      const cardType: CardType = cardTypeInt;
       const itemConfigCard = itemConfig.GetCard(cardType);
       if (itemConfigCard !== undefined) {
         this.moddedCardTypesArray.push(cardType);